extracted string to color function into helper folder
The string to colour function was moved to facilitate easier
navigation of the graph container. Additionally since it is
only a function and not a react FC it could be imoprted into
the three components that use it instead of passed down as
props. Doing this added some simplicity to the types in the props