Exit example implemented
Clicking example now populates cards that are not wrapped in
a link component. This allows the app to work the traditional
way and with our example button. useNavigate hook was used in
appplication card component to facilitate. Users can now also
click exit example button to reset all states and use the app
the traditional way.