Delete card functionality refactor
Cards are now deleted without the ref hook
I believe the delRef hook was being used to
prevent an issue with event bubbling. Becuase
the delete button is nested inside of the
Card component and card has an onClick the
prior teams were calling the onClick on the
card as well as the delete button.
event.stopPropagation was used to prevent
the bubbling the del ref hook is now able
to be removed