Health metrics function rewrite
The system information coming from si was strangley
not being passed to through the function. The function
was being called syncronously and because all the si
function are async the return at the end of the function
was always passing an empty array. This is because although
the data from si was coming in, it was being pushed into an
empty array defined at the top of the file inside of a .then
and becuase the return was executing synchrounously the array
was never being updated.

Added awaits to each of the si functions and now the data
is moving again