Current metrics now used to determine how new metrics are added to db
Prior to this change new metrics were added based on the an arbitrary
length variable that was changed after the awaiting a find call to
the mongo db. Current metrics is now queried directly. Two layers of
awaits are still necessary when calling the find on the metrics db
due to the queueing of all the await calls by the different services.
Am looking in to making it more streamm line