Grafana api key section commented out temporarily
Upon load of the graphscontainer component an event
call is made to electron (health request) which calls
the mongo fetch function which grabs all the health data
stored inside of the individual services databases. Inside
of the mongoFetch function there is a call for the grafana
api key database. In the microservices portion of the app
grafana is not used so there is no key. This was causing an
error and preventing data flow. This portion is being commented
out temporarily so that more specific controls can be made
to differentiate between apps that use docker (and thus grafana)
and microservices