The targeted keys state is now set from the rows function
prior to this change the targeted keys were populated from
the metrics pool state. This created an issue where the
rows in the table were not matching up with the values in
the metrics pool state. Now when the rows are selected
the rows variable is referenced so that the data is
the same